refactor(api): move recipe data types to basalt-api and invert recipes dep
basalt-api now owns RecipeId, Recipe, OwnedShapedRecipe, and
OwnedShapelessRecipe. basalt-recipes depends on basalt-api and provides
RecipeRegistryHandle impl for its concrete RecipeRegistry.

Eliminates basalt-recipes from basalt-api's main dependencies — the
last remaining blocker for cargo publish -p basalt-api. Adds version
specifiers to workspace deps for basalt-types, basalt-derive, and
basalt-protocol so cargo publish can resolve them.

basalt-recipes re-exports the moved types so existing
basalt_recipes::Recipe paths in basalt-server keep working without
churn.
