+++ crates/basalt-api/Cargo.toml
@@ -20,5 +20,4 @@ testing = []
 [dependencies]
 basalt-protocol = { workspace = true, optional = true }
 basalt-types = { workspace = true }
-basalt-recipes = { workspace = true }
 log = { workspace = true }
