+++ crates/basalt-api/src/recipes/id.rs
